+++ Spaxel/src/code/engine/LoadingScreen.java
@@ -31,7 +31,7 @@ public class LoadingScreen {
             e.printStackTrace();
         }
         overlay = new Sprite(1280, 720,1, 0xff000000);
-        progress = new UIBar(320,640,640, Math.PI/2, new Sprite(1,8,2, 0xffffffff));
+        progress = new UIBar(320,640,640, (float)Math.PI/2, new Sprite(1,8,2, 0xffffffff));
         message = new Label(640,680, "",font , 16f);
         buffer = new RenderBuffer(Game.GAME_WIDTH, Game.GAME_HEIGHT);
         title = new Label(640, 320, "SPAXEL", font, 128f);
