+++ Spaxel/src/code/entity/SpaceCarrier.java
@@ -10,9 +10,9 @@ import code.graphics.Sprite;
 public class SpaceCarrier extends Actor {
     private Sprite turret;
     private AnimSprite doors;
-    private double turretRot;
+    private float turretRot;
 
-    public SpaceCarrier(double x, double y, double rot, int health, Sprite sprite, Sprite turret, AnimSprite doors, double maxspeed, double acc) {
+    public SpaceCarrier(float x, float y, float rot, int health, Sprite sprite, Sprite turret, AnimSprite doors, float maxspeed, float acc) {
         super(x, y, rot, health, sprite, maxspeed, acc);
         this.turret = turret;
         this.doors = doors;
@@ -29,10 +29,10 @@ public class SpaceCarrier extends Actor {
     }
 
     public void render(int xPos, int yPos, RenderBuffer render){
-        double dx1 = Math.cos(rot+Math.PI/2);
-        double dy1 = Math.sin(rot+Math.PI/2);
-        double dx2 = Math.cos(rot+Math.PI);
-        double dy2 = Math.sin(rot+Math.PI);
+        float dx1 = (float)Math.cos(rot+Math.PI/2);
+        float dy1 = (float)Math.sin(rot+Math.PI/2);
+        float dx2 = (float)Math.cos(rot+Math.PI);
+        float dy2 = (float)Math.sin(rot+Math.PI);
         Sprite test =new Sprite(1,1,2,0xffff0000);
         sprite.render((int) (x + xPos), (int) (y + yPos), rot, render);
         turret.render((int) (x + dx1*4*5 + dx2*4*21 + xPos), (int) (y + dy1*4*5 + dy2*4*21 + yPos), rot+turretRot, render);
