+++ Spaxel/src/code/entity/TrailSegment.java
@@ -11,9 +11,9 @@ import javafx.scene.control.TreeItem;
 public class TrailSegment extends Entity{
     private Sprite trail;
     private TrailSegment previous;
-    private double maxLife;
+    private float maxLife;
 
-    public TrailSegment(double x, double y, double rot, Sprite trail, TrailSegment previous){
+    public TrailSegment(float x, float y, float rot, Sprite trail, TrailSegment previous){
         super(x,y, rot);
         this.trail = trail;
         this.previous = previous;
@@ -28,27 +28,27 @@ public class TrailSegment extends Entity{
         if (previous != null){
             int steps = 4;
             if (rot == previous.getRot()){
-                double xStep = (x - previous.getX())/steps;
-                double yStep = (y - previous.getY())/steps;
+                float xStep = (x - previous.getX())/steps;
+                float yStep = (y - previous.getY())/steps;
                 for (int i = 0; i < steps; i++){
                     trail.render((int)(previous.getX() + i*xStep) +xPos,(int)(previous.getY() + i*yStep)+yPos, render,1 - (life/maxLife)/2);
                 }
             }
             else {
-                double dx1 = Math.sin(rot+Math.PI);
-                double dy1 = Math.cos(rot+Math.PI);
-                double dx2 = Math.sin(previous.getRot());
-                double dy2 = Math.cos(previous.getRot());
-                double j  = (y*dx1 - previous.getY()*dx1 + previous.getX()*dy1 - x*dy1)/(dx1*dy2 - dx2*dy1);
-                double xMid = previous.getX() + j*dx2;
-                double yMid = previous.getY() + j*dy2;
-                double x1Step = (xMid - previous.getX())/steps;
-                double y1Step = (yMid - previous.getY())/steps;
-                double x2Step = (x - xMid)/steps;
-                double y2Step = (y - yMid)/steps;
+                float dx1 = (float)Math.sin(rot+Math.PI);
+                float dy1 = (float)Math.cos(rot+Math.PI);
+                float dx2 = (float)Math.sin(previous.getRot());
+                float dy2 = (float)Math.cos(previous.getRot());
+                float j  = (y*dx1 - previous.getY()*dx1 + previous.getX()*dy1 - x*dy1)/(dx1*dy2 - dx2*dy1);
+                float xMid = previous.getX() + j*dx2;
+                float yMid = previous.getY() + j*dy2;
+                float x1Step = (xMid - previous.getX())/steps;
+                float y1Step = (yMid - previous.getY())/steps;
+                float x2Step = (x - xMid)/steps;
+                float y2Step = (y - yMid)/steps;
                 for (int i = 1; i <= steps; i++){
-                    double xStep = ((xMid + i*x2Step) - (previous.getX() + i*x1Step))/steps;
-                    double yStep = ((yMid + i*y2Step) - (previous.getY() + i*y1Step))/steps;
+                    float xStep = ((xMid + i*x2Step) - (previous.getX() + i*x1Step))/steps;
+                    float yStep = ((yMid + i*y2Step) - (previous.getY() + i*y1Step))/steps;
                     trail.render((int)(previous.getX() + i*x1Step + i*xStep) +xPos,(int)(previous.getY() + i*y1Step + i*yStep)+yPos, render,1 - (life/maxLife)/2);
                 }
             }
