+++ Spaxel/src/code/factories/ClusterMissileFactory.java
@@ -9,11 +9,11 @@ import code.projectiles.Projectile;
  * Created by theo on 9-5-2016.
  */
 public class ClusterMissileFactory extends ProjectileFactory {
-    public ClusterMissileFactory(Sprite sprite, Sprite trail, int damage, int life, double speed) {
+    public ClusterMissileFactory(Sprite sprite, Sprite trail, int damage, int life, float speed) {
         super(sprite,trail, damage, life, speed);
     }
 
-    public Projectile make(double x, double y, double rot){
+    public Projectile make(float x, float y, float rot){
         return new ClusterMissile(x, y, rot, sprite,trail, damage,life, speed);
     }
 }
