+++ Spaxel/src/code/factories/DisruptLaserFactory.java
@@ -8,11 +8,11 @@ import code.projectiles.Projectile;
  * Created by theo on 12-5-2016.
  */
 public class DisruptLaserFactory extends ProjectileFactory{
-    public DisruptLaserFactory(Sprite sprite, Sprite trail, int damage, int life, double speed) {
+    public DisruptLaserFactory(Sprite sprite, Sprite trail, int damage, int life, float speed) {
         super(sprite,trail,  damage, life, speed);
     }
 
-    public Projectile make(double x, double y, double rot){
+    public Projectile make(float x, float y, float rot){
         return new DisruptLaser(x, y, rot, sprite,trail, damage, life, speed);
     }
 }
