+++ Spaxel/src/code/factories/PiercingLaserFactory.java
@@ -8,11 +8,11 @@ import code.projectiles.Projectile;
  * Created by theo on 11-5-2016.
  */
 public class PiercingLaserFactory extends ProjectileFactory {
-    public PiercingLaserFactory(Sprite sprite, Sprite trail, int damage, int life, double speed) {
+    public PiercingLaserFactory(Sprite sprite, Sprite trail, int damage, int life, float speed) {
         super(sprite,trail,  damage, life, speed);
     }
 
-    public Projectile make(double x, double y, double rot){
+    public Projectile make(float x, float y, float rot){
         return new PiercingLaser(x, y, rot, sprite,trail, damage, life, speed);
     }
 }
